Add material library integration and GeoScript support for PhysicoChem
Material Library Integration:
- Add compound/mineral selector UI to PhysicoChemTools with CompoundLibrary integration
- Search and filter minerals from thermodynamic database
- Select multiple minerals with adjustable volume fractions
- Auto-normalization of mineral fractions to sum to 1.0
- Material properties now include MineralComposition and MineralFractions fields
- Seamless integration with existing thermodynamic solvers

GeoScript PhysicoChem Commands:
- CREATE_REACTOR: Create new reactor with box domain programmatically
- ADD_DOMAIN: Add Box, Sphere, or Cylinder domains with position and size
- SET_MINERALS: Configure mineral composition with fractions for domains
- RUN_SIMULATION: Execute full multiphysics simulation with configurable timesteps
- All commands registered in GeoScript CommandRegistry

Enhancements:
- Fix type conversion errors in PhysicoChemDataset ImportFromDTO for Profile2D and CustomPoints
- Material selector UI with search filter and checkbox selection
- Fraction sliders with visual feedback (green for normalized, orange for warning)
- Example workflow: CREATE_REACTOR reactor 2 2 2 |> SET_MINERALS Calcite 0.7 Quartz 0.3 |> RUN_SIMULATION 100 0.1

This enables easy reactor setup through both UI and scripting

+++ Data/PhysicoChem/PhysicoChemDataset.cs
@@ -414,10 +414,10 @@ public class PhysicoChemDataset : Dataset, ISerializableDataset
                     Radius = domainDto.Geometry.Radius,
                     InnerRadius = domainDto.Geometry.InnerRadius,
                     Height = domainDto.Geometry.Height,
-                    Profile2D = domainDto.Geometry.Profile2D?.Select(p => (p.X, p.Y)).ToList() ?? new List<(double, double)>(),
+                    Profile2D = domainDto.Geometry.Profile2D?.Select(p => ((double)p.X, (double)p.Y)).ToList() ?? new List<(double, double)>(),
                     ExtrusionDepth = domainDto.Geometry.ExtrusionDepth,
                     RadialSegments = domainDto.Geometry.RadialSegments,
-                    CustomPoints = domainDto.Geometry.CustomPoints?.Select(p => (p.X, p.Y, p.Z)).ToList() ?? new List<(double, double, double)>(),
+                    CustomPoints = domainDto.Geometry.CustomPoints?.Select(p => ((double)p.X, (double)p.Y, (double)p.Z)).ToList() ?? new List<(double, double, double)>(),
                     MeshFilePath = domainDto.Geometry.MeshFilePath
                 };
             }
